Imported upstream version '2.36.0' of 'upstream'

+++ transmission_interface/test/random_generator_utils.hpp
@@ -31,7 +31,7 @@ struct RandomDoubleGenerator
 public:
   RandomDoubleGenerator(double min_val, double max_val) : min_val_(min_val), max_val_(max_val)
   {
-    srand(static_cast<unsigned int>(time(nullptr)));
+    srand(time(nullptr));
   }
   double operator()()
   {
